chore: merge bvs-test-utils into bvs-testing
