+++ crates/Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    # production contracts
     "bvs-delegation-manager",
     "bvs-directory",
     "bvs-driver",
@@ -11,7 +12,6 @@ members = [
     "bvs-strategy-factory",
     "bvs-strategy-manager",
     "bvs-test",
-    "common/bvs/test-utils",
 ]
 
 resolver = "2"
@@ -61,5 +61,4 @@ bvs-strategy-manager = { path = "./bvs-strategy-manager" }
 
 # common cosmwasm crates
 bvs-base = { path = "./common/bvs/base" }
-bvs-test-utils = { path = "./common/bvs/test-utils" }
 bvs-testing = { path = "./common/bvs/testing" }
