+++ crates/bvs-slash-manager/Cargo.toml
@@ -36,4 +36,4 @@ serde_json = { workspace = true }
 
 [dev-dependencies]
 cw-multi-test = { workspace = true }
-bvs-test-utils = { workspace = true }
+bvs-testing = { workspace = true }
