--- crates/common/bvs/test-utils/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "bvs-test-utils"
-description = "Test utility for bvs contracts"
-version = "0.0.0"
-edition.workspace = true
-
-[dependencies]
-cosmwasm-std = { workspace = true }
-bech32 = { workspace = true }
-ripemd = { workspace = true }
-sha2 = { workspace = true }
-secp256k1 = { workspace = true }
-cosmwasm-crypto = { workspace = true }
