+++ crates/common/bvs/testing/Cargo.toml
@@ -11,6 +11,15 @@ documentation = { workspace = true }
 keywords = { workspace = true }
 
 [dependencies]
+cosmwasm-std = { workspace = true }
+cw-multi-test = { workspace = true }
+bech32 = { workspace = true }
+ripemd = { workspace = true }
+serde = { workspace = true }
+sha2 = { workspace = true }
+secp256k1 = { workspace = true }
+cosmwasm-crypto = { workspace = true }
+
 bvs-delegation-manager = { workspace = true }
 bvs-directory = { workspace = true }
 bvs-driver = { workspace = true }
@@ -21,5 +30,3 @@ bvs-strategy-base = { workspace = true }
 bvs-strategy-base-tvl-limits = { workspace = true }
 bvs-strategy-factory = { workspace = true }
 bvs-strategy-manager = { workspace = true }
-cosmwasm-std = { workspace = true }
-cw-multi-test = { workspace = true }
