+++ crates/common/bvs/testing/src/bvs_directory_querier.rs
@@ -1,3 +1,4 @@
+#![allow(dead_code)]
 use std::collections::HashMap;
 
 use bvs_delegation_manager::msg::QueryMsg;
