+++ crates/common/bvs/testing/src/lib.rs
@@ -1,9 +1,10 @@
+pub mod account;
 pub mod integration;
 
-mod bvs_mock_querier;
-mod mocks;
 mod bvs_delegation_manager_querier;
 mod bvs_directory_querier;
+mod bvs_mock_querier;
+mod mocks;
 
 pub use bvs_mock_querier::BvsMockQuerier;
 pub use mocks::*;
